Fix unicode processing
Now, instead of processing strings, we work with bytes, which does not lead to an exception when processing some unicode characters, such as emoji.

Also, now there is no decoding from utf-8 and vice versa, which can potentially increase performance.

The application code has been moved to a function, so it is easier to debug and this is a more pythonic way.